Added @STREAMobj/intersectlocs.m
intersectlocs returns for two stream networks S and S1 (STREAMobj) derived from a FLOWobj the locations where S1 starts to have common flowpath with S. The function may be used to map confluences of a low order stream network with a  network that contains only higher order streams. It may also be useful to map intersections of landslide runout zones with a river network.